refector: class로 변경
