+++ src/js/Components/CalculateTotal.js
@@ -1,15 +1,15 @@
-export default function CalculateTotal({ $total, state }) {
-  this.state = state;
-  this.$total = $total;
-
-  this.render = () => {
-    this.$total.textContent = this.state.total || 0;
-  };
-
-  this.setState = (nextState) => {
-    this.state = nextState;
+import component from '../core/component.js';
+export default class CalculateTotal extends component {
+  constructor($target, state) {
+    super($target, state);
     this.render();
+  }
+
+  render = () => {
+    this.$target.textContent = this.state.total || '0';
   };
 
-  this.render();
+  setState = (newState) => {
+    super.setState(newState);
+  };
 }
