+++ src/js/index.js
@@ -1,8 +1,6 @@
-import App from './App.js';
+import App from './Components/App.js';
 
+const $app = document.querySelector('#app');
 const initialState = { total: '', digitCount: 0, operation: '' };
 
-new App({
-  $app: document.querySelector('#app'),
-  initialState: initialState || undefined,
-});
+new App($app, initialState || undefined);
